some style on quiz
