+++ src/quiz/animal-totem.scss
@@ -24,3 +24,10 @@
     margin:5px;
     font-weight: bold;
 }
+
+@media screen and (max-width: 768px){
+    .animal-wrapper{
+        width:94%
+      }
+ 
+}
