+++ src/quiz/quiz.scss
@@ -29,6 +29,7 @@
         font-size: 1.6em;
         margin-bottom: 1em;
         text-align: center;
+       
       }
  
 }
